* Fixed read callback
git-svn-id: svn://svn.code.sf.net/p/gmerlin/code/trunk@1379 214d8ac5-8513-4f59-80f8-2f39b674a24a

+++ gmerlin_avdecoder/lib/audio_flac.c
@@ -57,7 +57,7 @@ typedef struct
 static FLAC__StreamDecoderReadStatus
 read_callback(const FLAC__StreamDecoder *decoder,
               FLAC__byte buffer[],
-              unsigned *bytes, void *client_data)
+              size_t *bytes, void *client_data)
   {
   int bytes_read = 0;
   int bytes_to_copy;
